+++ ReleaseNotes.md
@@ -6,6 +6,7 @@
 - Add Information Protection as new collector (Only working when running with app registration)
 - Bugfix #4 Connect with Secure String
 - Add IdentityProvider.Read.All to scopes as it is required by the AzureAD part
+- Add Windows 365 as new collector (Beta)
 
 ## 3.0.3 - 26.05.2021 - Thomas Kurth